Skip failing test
devtools::test() only runs test files that start with "test", so this effectively skips the test. I ported this test over from the old inst/tests/ directory, but most of it fails. If we want, we can look into fixing it, but for now, just copy it over and keep the test source around.