Complete removal of inst/tests/
After researching the difference between inst/tests/ and tests/, I conclude that we don't need to put tests in inst/tests/. The main difference is inst/tests/ is included in installations of the package, while tests/ is excluded. I don't think people installing the package need the tests, so I think the default testthis location is fine (tests/).